backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
faiss-cpu==1.13.2
openai==1.61.0
langchain==0.3.19
langchain-openai==0.3.7
langgraph==0.2.74
sentence-transformers==3.0.0
elevenlabs==1.0.0
python-multipart==0.0.20
httpx==0.28.1
orjson==3.10.15
scikit-learn==1.6.1
